participant routes changed

+++ src/functions/participant/skills/addSkills.js
@@ -1,10 +1,18 @@
 const Skill = require('../../../models/Skill')
-//transaction control
-//check if incoming skills = []
+
 const addSkills = async(req,res) => {
     try {
-        await Skill.deleteMany({participant_id:req.participant._id})
+    let check = 0
     const skills = req.body.skills
+    if(skills.length==0){
+        return res.send('Please enter some skills')
+    }
+    const allowed_skills = ['frontend','backend','ml','ui/ux','appdev']
+    const isAllowed = skills.every((skill)=>{allowed_skills.includes(skill)})
+    if(!isAllowed){
+        return res.send('Invalid skills')
+    }
+    await Skill.deleteMany({participant_id:req.participant._id})
     let skillRecords = []
     let i = 0
     skills.forEach(async(skill) => {
@@ -14,8 +22,12 @@ const addSkills = async(req,res) => {
         })
         await newSkill.save()
         skillRecords.push(newSkill)
+        i++
+        if(i==skills.length){
+            res.status(201).send(skillRecords)
+        }
     })
-    res.status(201).send(skillRecords)
+    
     } catch (e) {
         res.status(400).send(e)   
     }
